# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "ALAHLIFC API",
        "version": "1.1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/comments": {
            "post": {
                "tags": [
                    "comment"
                ],
                "summary": "get all comments",
                "operationId": "comments",
                "parameters": [
                    {
                        "name": "access-token",
                        "in": "header",
                        "type": "string"
                    },
                    {
                        "name": "link",
                        "in": "formData",
                        "description": "this link type only",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "type",
                        "in": "formData",
                        "description": "default type is video and can use (news,video,product)",
                        "type": "string"
                    },
                    {
                        "name": "limit",
                        "in": "formData",
                        "description": "default limit 12  and can you increase limit",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "error"
                    }
                }
            }
        },
...
    },
...
}